pyproject.toml
[project]
name = "backup-pilot"
version = "0.1.0"
description = "A cross-platform CLI utility for backing up and restoring multiple databases with pluggable storage backends."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "BackupPilot Authors" }
]
keywords = ["backup", "database", "cli", "mysql", "postgresql", "mongodb", "sqlite"]
dependencies = [
  "typer[all]>=0.12.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "boto3>=1.34.0",
  "google-cloud-storage>=2.16.0",
  "azure-storage-blob>=12.20.0",
  "requests>=2.32.0",
  "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "black>=24.0.0",
  "ruff>=0.4.0",
  "mypy>=1.9.0",
]

[project.scripts]
"backup-pilot" = "backup_pilot.cli.main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backup_pilot*"]
exclude = ["backups*"]

